fix: architecture issues
