+++ studio/lib/interfaces/socialMedia.ts
@@ -1,5 +1,3 @@
-import { soMeLinksID } from "studio/schemas/documents/siteSettings/socialMediaProfiles";
-
 // Define the SocialMediaLink interface based on the expected structure
 export interface SocialMediaLink {
   _type: string;
@@ -11,6 +9,6 @@ export interface SocialMediaLink {
 // Define the SocialMediaProfiles interface based on the document schema
 export interface SocialMediaProfiles {
   _id: string;
-  _type: typeof soMeLinksID;
+  _type: "soMeLinksID";
   soMeLinkArray?: SocialMediaLink[];
 }
