+++ studioShared/lib/interfaces/splitSection.ts
@@ -1,6 +1,3 @@
-import { isSanityKeyTypeObject } from "studio/lib/interfaces/global";
-import { splitSectionSections } from "studioShared/schemas/objects/splitSection";
-
 import { ImageBlock } from "./imageBlock";
 import { TextBlock } from "./textBlock";
 
@@ -15,17 +12,4 @@ export interface SplitSection {
   _key: string;
   _type: "splitSection";
   sections: SplitSectionSection[];
-}
-
-export function isSplitSectionSections(
-  value: unknown,
-): value is SplitSectionSection[] {
-  return (
-    Array.isArray(value) &&
-    value.every(
-      (item) =>
-        isSanityKeyTypeObject(item) &&
-        splitSectionSections.some((s) => s.name === item._type),
-    )
-  );
 }
